Implement extra contrast changes in addition to the original contrast change. Now supports up to 3 separate contrast changes (and times at which they should occur).

+++ src/fixstim/AttentionStimSet.cpp
@@ -425,7 +425,7 @@ AttentionStimSet::AttentionStimSet(ARContrastFixationPointSpec& fixpt, double tM
 int AttentionStimSet::init(ARvsg& vsg, std::vector<int> pages)
 {
 	int status = 0;
-	int nlevels;
+	size_t nlevels;
 	m_pageBlank = pages[0];
 	m_pageFixpt = pages[1];
 	m_pageStim = pages[2];
